refactor: move up events analytics to new file track_up
