+++ pkg/analytics/track.go
@@ -32,15 +32,8 @@ import (
 const (
 	// skipcq GSC-G101
 	// This is mixpanel's public token, is needed to send analytics to the project
-	mixpanelToken = "92fe782cdffa212d8f03861fbf1ea301"
-
-	upEvent                  = "Up"
-	upErrorEvent             = "Up Error"
+	mixpanelToken            = "92fe782cdffa212d8f03861fbf1ea301"
 	manifestHasChangedEvent  = "Manifest Has Changed"
-	durationActivateUpEvent  = "Up Duration Time"
-	reconnectEvent           = "Reconnect"
-	durationInitialSyncEvent = "Initial Sync Duration Time"
-	syncErrorEvent           = "Sync Error"
 	syncResetDatabase        = "Sync Reset Database"
 	downEvent                = "Down"
 	downVolumesEvent         = "DownVolumes"
@@ -133,87 +126,14 @@ func TrackPreviewDestroy(success bool) {
 	track(previewDestroyEvent, success, nil)
 }
 
-const (
-	// ReconnectCauseDefault is the default cause for a reconnection
-	ReconnectCauseDefault = "unrecognised"
-
-	// ReconnectCauseDevPodRecreated is cause when pods UID change between retrys
-	ReconnectCauseDevPodRecreated = "dev-pod-recreated"
-)
-
-// TrackReconnect sends a tracking event to mixpanel when the development container reconnect
-func TrackReconnect(success bool, cause string) {
-	props := map[string]interface{}{
-		"cause": cause,
-	}
-	track(reconnectEvent, success, props)
-}
-
-// TrackSyncError sends a tracking event to mixpanel when the init sync fails
-func TrackSyncError() {
-	track(syncErrorEvent, false, nil)
-}
-
-// TrackDurationInitialSync sends a tracking event to mixpanel with initial sync duration
-func TrackDurationInitialSync(durationInitialSync time.Duration) {
-	props := map[string]interface{}{
-		"duration": durationInitialSync,
-	}
-	track(durationInitialSyncEvent, true, props)
-}
-
 // TrackResetDatabase sends a tracking event to mixpanel when the syncthing database is reset
 func TrackResetDatabase(success bool) {
 	track(syncResetDatabase, success, nil)
 }
 
-// TrackUpMetadata defines the properties an up can have
-type TrackUpMetadata struct {
-	IsV2                   bool
-	ManifestType           model.Archetype
-	IsInteractive          bool
-	IsOktetoRepository     bool
-	HasDependenciesSection bool
-	HasBuildSection        bool
-	HasDeploySection       bool
-	Success                bool
-	HasReverse             bool
-	IsHybridDev            bool
-	Mode                   string
-}
-
-// TrackUp sends a tracking event to mixpanel when the user activates a development container
-func TrackUp(m TrackUpMetadata) {
-	props := map[string]interface{}{
-		"isInteractive":          m.IsInteractive,
-		"isV2":                   m.IsV2,
-		"manifestType":           m.ManifestType,
-		"isOktetoRepository":     m.IsOktetoRepository,
-		"hasDependenciesSection": m.HasDependenciesSection,
-		"hasBuildSection":        m.HasBuildSection,
-		"hasDeploySection":       m.HasDeploySection,
-		"hasReverse":             m.HasReverse,
-		"mode":                   m.Mode,
-	}
-	track(upEvent, m.Success, props)
-}
-
-// TrackUpError sends a tracking event to mixpanel when the okteto up command fails
-func TrackUpError(success bool) {
-	track(upErrorEvent, success, nil)
-}
-
 // TrackManifestHasChanged sends a tracking event to mixpanel when the okteto up command fails because manifest has changed
 func TrackManifestHasChanged(success bool) {
 	track(manifestHasChangedEvent, success, nil)
-}
-
-// TrackDurationActivateUp sends a tracking event to mixpanel of the time that has elapsed in the execution of up
-func TrackDurationActivateUp(durationActivateUp time.Duration) {
-	props := map[string]interface{}{
-		"duration": durationActivateUp,
-	}
-	track(durationActivateUpEvent, true, props)
 }
 
 // TrackExecMetadata is the metadata added to execEvent
